add debug flag for setup_chroot (Install-Common.sh)

+++ extra/ISO/Install-Common.sh
@@ -765,24 +765,26 @@ setup_grub_theme() {
 
 # NixOS doesn't use this
 setup_chroot() {
-  # Run Setup Script
-  local distro="${1:-}"
-  local MNT="/mnt"
-  local CHROOT="chroot"
-  # Gentoo uses /mnt/gentoo instead of /mnt
-  [ "$distro" = "gentoo" ] && MNT="/mnt/gentoo"
-  # Prefer xchroot, then arch-chroot, fall back to chroot
-  command -v xchroot >/dev/null 2>&1 && CHROOT="xchroot"
-  command -v arch-chroot >/dev/null 2>&1 && 
-    [ "$distro" != "slackware" ] && CHROOT="arch-chroot"
-    echo "Basic install complete. Running post-install setup..."
-  $CHROOT "$MNT" su - "$username" -c "
-    cd ~/cinnamon-dotfiles &&
-    SUDO_PASSWORD='$userpasswd' bash Setup.sh
-  " || die "Post-install setup failed."
-  # Place cinnamon-ISO flag if successful
-  $CHROOT "$MNT" su - "$username" \
-    -c "touch ~/cinnamon-dotfiles/.iso.done"
+  if [[ ! -f ".debug" ]]; then
+    # Run Setup Script
+    local distro="${1:-}"
+    local MNT="/mnt"
+    local CHROOT="chroot"
+    # Gentoo uses /mnt/gentoo instead of /mnt
+    [ "$distro" = "gentoo" ] && MNT="/mnt/gentoo"
+    # Prefer xchroot, then arch-chroot, fall back to chroot
+    command -v xchroot >/dev/null 2>&1 && CHROOT="xchroot"
+    command -v arch-chroot >/dev/null 2>&1 && 
+      [ "$distro" != "slackware" ] && CHROOT="arch-chroot"
+      echo "Basic install complete. Running post-install setup..."
+    $CHROOT "$MNT" su - "$username" -c "
+      cd ~/cinnamon-dotfiles &&
+      SUDO_PASSWORD='$userpasswd' bash Setup.sh
+    " || die "Post-install setup failed."
+    # Place cinnamon-ISO flag if successful
+    $CHROOT "$MNT" su - "$username" \
+      -c "touch ~/cinnamon-dotfiles/.iso.done"
+  fi
 }
 
 # Only Void uses this
